+++ common/src/main/java/dev/qixils/crowdcontrol/common/command/impl/maxhealth/MaxHealthSubCommand.java
@@ -24,7 +24,7 @@ public class MaxHealthSubCommand<P> implements ImmediateCommand<P> {
 	public @NotNull Component getProcessedDisplayName(@NotNull Request request) {
 		if (request.getParameters() == null)
 			return getDefaultDisplayName();
-		return getDefaultDisplayName().args(Component.text((int) request.getParameters()[0]));
+		return getDefaultDisplayName().args(Component.text((int) (double) request.getParameters()[0]));
 	}
 
 	@NotNull
